+++ source/refined-github.css
@@ -109,7 +109,7 @@
 }
 
 /* Improve spacing in discussion lists meta, made feel crowded by:
-highlight-collaborators-and-own-discussions
+highlight-collaborators-and-own-conversations
 parse-backticks
 pr-branches
 */
