+++ source/refined-github.ts
@@ -13,7 +13,7 @@ import './features/hide-tips.css';
 import './features/hide-readme-header.css';
 import './features/hide-obvious-tooltips.css';
 import './features/clean-discussions.css';
-import './features/sticky-discussion-list-toolbar.css';
+import './features/sticky-conversation-list-toolbar.css';
 import './features/always-show-branch-delete-buttons.css';
 import './features/easier-pr-sha-copy.css';
 import './features/repo-stats-spacing.css';
@@ -60,9 +60,9 @@ import './features/batch-open-issues';
 import './features/hide-useless-comments';
 import './features/navigate-pages-with-arrow-keys';
 import './features/discussion-links-on-repo-lists';
-import './features/global-discussion-list-filters';
-import './features/discussion-filters';
-import './features/sort-issues-by-update-time'; // Must be after global-discussion-list-filters and discussion-filters and discussion-links-on-repo-lists
+import './features/global-conversation-list-filters';
+import './features/conversation-filters';
+import './features/sort-issues-by-update-time'; // Must be after global-conversation-list-filters and conversation-filters and discussion-links-on-repo-lists
 import './features/pinned-issues-update-time';
 import './features/latest-tag-button';
 import './features/default-branch-button';
@@ -144,7 +144,7 @@ import './features/restore-file';
 import './features/hidden-review-comments-indicator';
 import './features/reload-failed-proxied-images';
 import './features/clean-rich-text-editor';
-import './features/highlight-collaborators-and-own-discussions';
+import './features/highlight-collaborators-and-own-conversations';
 import './features/embed-gist-via-iframe';
 import './features/expand-all-collapsed-code';
 import './features/separate-draft-pr-button';
@@ -157,7 +157,7 @@ import './features/linkify-user-location';
 import './features/repo-age';
 import './features/user-local-time';
 import './features/quick-mention';
-import './features/extend-discussion-status-filters';
+import './features/extend-conversation-status-filters';
 import './features/expand-all-hidden-comments';
 import './features/bugs-tab';
 import './features/cross-deleted-pr-branches';
